requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiohttp-retry==2.9.1
aiosignal==1.4.0
asgiref==3.8.1
attrs==25.3.0
certifi==2025.6.15
charset-normalizer==3.4.2
diff-match-patch==20241021
Django==5.2.3
django-import-export==4.3.8
docopt==0.6.2
frozenlist==1.7.0
idna==3.10
multidict==6.6.3
narwhals==1.45.0
num2words==0.5.14
packaging==25.0
pillow==11.2.1
plotly==6.2.0
propcache==0.3.2
pycountry==24.6.1
PyJWT==2.10.1
reportlab==4.4.2
requests==2.32.4
sqlparse==0.5.3
tablib==3.8.0
twilio==9.6.4
typing_extensions==4.14.0
urllib3==2.5.0
yarl==1.20.1
